rename localStorage folder to storage
