+++ application/ui/src/features/dataset/import-export/export-jobs-list/export-job/export-active-job.component.tsx
@@ -3,7 +3,7 @@
 
 import { Divider, Flex, Loading, Text, View } from '@geti/ui';
 import { getJobProgress, isJobRunning } from 'hooks/api/util';
-import { useExportDataset } from 'hooks/localStorage/use-export-dataset.hook';
+import { useExportDataset } from 'hooks/storage/use-export-dataset.hook';
 
 import { ExportDatasetJob } from '../../../../../constants/shared-types';
 import { BottomProgressBar } from '../../../../models/model-listing/current-model-running/bottom-progress-bar.component';
