+++ application/ui/src/features/dataset/import-export/export-jobs-list/export-job/export-completed-job/export-completed-job.test.tsx
@@ -24,7 +24,7 @@ vi.mock('../../../../../../shared/util', async (importActual) => {
 
 const mockedRemoveLsExportId = vi.fn();
 
-vi.mock('../../../../../../hooks/localStorage/use-export-dataset.hook', () => ({
+vi.mock('../../../../../../hooks/storage/use-export-dataset.hook', () => ({
     useExportDataset: () => ({
         removeLsExportId: mockedRemoveLsExportId,
         getLsExportIds: vi.fn(() => []),
