+++ application/ui/src/features/dataset/import-export/export-jobs-list/hooks/use-export-status.hook.test.tsx
@@ -11,7 +11,7 @@ import { useExportStatus } from './use-export-status.hook';
 
 const mockRemoveLsExportId = vi.fn();
 
-vi.mock('hooks/localStorage/use-export-dataset.hook', () => ({
+vi.mock('hooks/storage/use-export-dataset.hook', () => ({
     useExportDataset: () => ({
         removeLsExportId: mockRemoveLsExportId,
         getLsExportIds: vi.fn(() => []),
