+++ application/ui/src/features/dataset/import-export/import-dataset/import-dataset-buttons/import-dataset-buttons.component.tsx
@@ -4,7 +4,7 @@
 import { Button, ButtonGroup } from '@geti/ui';
 
 import { ImportJobProcessButtons } from '../../../../../components/import-job-process/import-job-process-buttons.component';
-import { useImportDatasetToProject } from '../../../../../hooks/localStorage/use-import-dataset-to-project.hook';
+import { useImportDatasetToProject } from '../../../../../hooks/storage/use-import-dataset-to-project.hook';
 import { LabelMappingButtons } from '../label-mapping/label-mapping-buttons.component';
 import { ImportDatasetToProjectState } from '../util';
