+++ application/ui/src/features/dataset/import-export/import-dataset/label-mapping/label-mapping.component.tsx
@@ -9,7 +9,7 @@ import { useStagedDataset } from 'hooks/api/staged-dataset.hook';
 import { $api } from '../../../../../api/client';
 import { DatasetStatistics } from '../../../../../components/dataset-statistics/dataset-statistics.component';
 import { useProject } from '../../../../../hooks/api/project.hook';
-import { useImportDatasetToProject } from '../../../../../hooks/localStorage/use-import-dataset-to-project.hook';
+import { useImportDatasetToProject } from '../../../../../hooks/storage/use-import-dataset-to-project.hook';
 import { useImportDatasetDialogState } from '../../../providers/export-import-dataset-dialog-provider.component';
 import { FormatWarning } from './format-warning/format-warning.component';
 import { IMPORT_DATASET_FORM_ID, mapProjectLabels } from './util';
