+++ application/ui/src/features/dataset/import-export/import-jobs-list/import-jobs-list.component.tsx
@@ -3,7 +3,7 @@
 
 import { Flex } from '@geti/ui';
 import { useQueryClient } from '@tanstack/react-query';
-import { useImportDatasetToProject } from 'hooks/localStorage/use-import-dataset-to-project.hook';
+import { useImportDatasetToProject } from 'hooks/storage/use-import-dataset-to-project.hook';
 import { useProjectIdentifier } from 'hooks/use-project-identifier.hook';
 import { isEmpty, partition } from 'lodash-es';
