+++ application/ui/src/features/project/list/import-dataset-as-new-project/import-label-mapping/import-label-mapping.component.tsx
@@ -5,7 +5,7 @@ import { useActionState } from 'react';
 
 import { Checkbox, dimensionValue, Divider, Flex, Form, Heading, View } from '@geti/ui';
 import { useStagedDataset } from 'hooks/api/staged-dataset.hook';
-import { useImportDatasetAsNewProject } from 'hooks/localStorage/use-import-dataset-as-new-project.hook';
+import { useImportDatasetAsNewProject } from 'hooks/storage/use-import-dataset-as-new-project.hook';
 import { isNil, isString } from 'lodash-es';
 
 import { $api } from '../../../../../api/client';
