+++ application/ui/src/features/project/list/import-dataset-as-new-project/import-process/import-process.component.tsx
@@ -1,7 +1,7 @@
 // Copyright (C) 2026 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { useImportDatasetAsNewProject } from 'hooks/localStorage/use-import-dataset-as-new-project.hook';
+import { useImportDatasetAsNewProject } from 'hooks/storage/use-import-dataset-as-new-project.hook';
 
 import { ImportJobProcess } from '../../../../../components/import-job-process/import-job-process.component';
 import { useImportDatasetDialog } from '../../../providers/import-dataset-dialog-provider.component';
